backend/pyproject.toml
[project]
name = "caretriage-api"
version = "0.1.0"
description = "CareTriage AI FastAPI backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "sqlalchemy>=2.0.36",
    "psycopg[binary]>=3.2.0",
    "PyJWT[crypto]>=2.10.0",
    "httpx>=0.28.0",
    "openai>=1.59.0",
    "python-multipart>=0.0.20",
]
